making everything private + CounterVisitor works on current command
