+++ src/test/java/edu/kis/powp/jobs2d/TestJobs2dApp.java
@@ -52,8 +52,8 @@ public class TestJobs2dApp {
 
 		application.addTest("Run command", new SelectRunCurrentCommandOptionListener(DriverFeature.getDriverManager()));
 		application.addTest("Deep secret command copy", new SelectDeepCopySecretCommand());
-		application.addTest("Visitor Command Counting test", new DriverCommandCounterVisitorTest());
-		application.addTest("Visitor Command Executor test", new DriverCommandExecutorVisitorTest(DriverFeature.getDriverManager().getCurrentDriver()));
+		application.addTest("Visitor Command Counting", new DriverCommandCounterVisitorTest());
+		application.addTest("Visitor Command Execution", new DriverCommandExecutorVisitorTest(DriverFeature.getDriverManager().getCurrentDriver()));
 
 		application.addTest("Rotate command 90 deg clockwise", new SelectCommandTransformationRotateOptionListener(90));
 		application.addTest("Rotate command 90 deg counterclockwise", new SelectCommandTransformationRotateOptionListener(-90));
