+++ src/test/java/edu/kis/powp/jobs2d/events/SelectDriverTransformation.java
@@ -8,7 +8,7 @@ import java.awt.event.ActionListener;
 
 public class SelectDriverTransformation implements ActionListener
 {
-	CoordsTransformation transformation;
+	private CoordsTransformation transformation;
 
 	public SelectDriverTransformation(CoordsTransformation transformation)
 	{
